Extract sensitive storage secure contract
